+++ packages/cli-v3/e2e/fixtures/esm-only-external/src/trigger/helloWorld.ts
@@ -1,0 +1,11 @@
+import { task } from "@trigger.dev/sdk/v3";
+import * as mupdf from "mupdf";
+
+export const helloWorld = task({
+  id: "helloWorld",
+  run: async () => {
+    console.log("Hello, World!", {
+      metaformat: mupdf.PDFDocument.META_FORMAT,
+    });
+  },
+});