+++ packages/core/src/v3/utils/ioSerialization.ts
@@ -160,6 +160,31 @@ export async function conditionallyImportPacket(
   }
 }
 
+export async function resolvePresignedPacketUrl(
+  url: string,
+  tracer?: TriggerTracer
+): Promise<any | undefined> {
+  try {
+    const response = await fetch(url);
+
+    if (!response.ok) {
+      return;
+    }
+
+    const data = await response.text();
+    const dataType = response.headers.get("content-type") ?? "application/json";
+
+    const packet = {
+      data,
+      dataType,
+    };
+
+    return await parsePacket(packet);
+  } catch (error) {
+    return;
+  }
+}
+
 async function importPacket(packet: IOPacket, span?: Span): Promise<IOPacket> {
   if (!packet.data) {
     return packet;
@@ -341,7 +366,18 @@ function getPacketExtension(outputType: string): string {
 }
 
 async function loadSuperJSON() {
-  return await import("superjson");
+  const superjson = await import("superjson");
+
+  superjson.registerCustom<Buffer, number[]>(
+    {
+      isApplicable: (v): v is Buffer => v instanceof Buffer,
+      serialize: (v) => [...v],
+      deserialize: (v) => Buffer.from(v),
+    },
+    "buffer"
+  );
+
+  return superjson;
 }
 
 function safeJsonParse(value: string): any {